+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 Next
 ----
 
+- Fix bug which meant that a user could not log in after ``dcos-docker wait`` on DC/OS Open Source clusters.
+
 2018.07.22.1
 ------------